+++ src/main/java/pl/playerony/repository/impl/RoleRepositoryImpl.java
@@ -5,12 +5,15 @@ import pl.playerony.exception.InputException;
 import pl.playerony.manager.SqlManager;
 import pl.playerony.model.impl.Role;
 import pl.playerony.repository.RoleRepository;
+import pl.playerony.util.SqlUtil;
 
 public class RoleRepositoryImpl implements RoleRepository {
 	private SqlManager sqlManager;
+	private SqlUtil sqlUtil;
 	
 	public RoleRepositoryImpl() {
 		sqlManager = new SqlManager();
+		sqlUtil = new SqlUtil();
 	}
 	
 	@Override
@@ -29,6 +32,9 @@ public class RoleRepositoryImpl implements RoleRepository {
 
 	@Override
 	public Role findRoleById(Long id) throws DatabaseException {
+		if(!sqlUtil.checkId("roles", id))
+			throw new DatabaseException("This id[" + id + "] dont exist in roles table");
+		
 		String sql = "SELECT * FROM roles "
 				   + "WHERE id = ?";
 		
@@ -41,6 +47,9 @@ public class RoleRepositoryImpl implements RoleRepository {
 
 	@Override
 	public Boolean removeRole(Long id) throws DatabaseException {
+		if(!sqlUtil.checkId("roles", id))
+			throw new DatabaseException("This id[" + id + "] dont exist in roles table");
+		
 		String sql = "DELETE FROM roles "
 				   + "	WHERE id = ?";
